My edits to this chapter took place at a few levels. I reduced the use of capitalization a lot, but I tried to be consistent. It is sort of strange to have "task" versus "ClusterTask" or "pipeline" versus "PipelineRun," but I want everyday words to look like everyday words (and hence limited capitalization).
I added some more of a hierarchy to the headings. There are more ## headings.

Following are questions I had about particular sentences.

---

After a while, Quarkus will generate and push your image to your registry. In my case, it’s `quay.io/wpernath/person-service?`.

I don't think you mean the question mark to be part of the name. It's not shown in Figure 3.

---

![Image 5: Pipeline Builder](Bildschirmfoto%202021-07-13%20um%2015.00.18.png)
![Image 6:Linking the workspaces from Pipeline to task](DraggedImage.png)

There's no indication what these figures are for. One shows a pipeline without parameters, and one shows parameters, but the figures don't fit in the sequence you're talking about.

---

You can log, list, cancel, and delete PipelineRuns.

This kind of sits in the middle of a different discussion, and it would be useful to say more about this possibilities.

---

Instead, I removed most of the parameters and added a MUST-HAVE maven settings ConfigMap.

What makes something "must-have." What does that phrase refer to in the ConfigMap? I have to admit that I have lost track in this paragraph of what you're doing. Why create a new file in the first place? Are you trying to solve just one problem (caching), or make it easier to customize builds in general?